GCCPY: 	* NEW fast and efficient Constant Folder for python 	* Cleanups of the passes and runtime code

+++ gcc/python/dot-hashtab.h
@@ -19,7 +19,7 @@ along with GCC; see the file COPYING3.  If not see
 
 typedef struct gpy_vector_t {
   void ** vector;
-  signed long size, length;
+  size_t size, length;
 } cm_vector_t;
 
 typedef unsigned long gpy_hashval_t;
@@ -43,11 +43,11 @@ typedef gpy_vector_t * dot_contextTable_t;
 #define NULL_VEC (gpy_vector_t*)0
 #define VEC_length(x)  x->length
 #define VEC_index(T,V,I)				\
-  (T)gpy_vec_index_diag (V,(int)I, __FILE__, __LINE__)
+  (T)gpy_vec_index_diag (V, I, __FILE__, __LINE__)
 
 extern void gpy_vec_push (gpy_vector_t * const, void *);
 extern void * gpy_vec_pop (gpy_vector_t * const);
-extern void * gpy_vec_index_diag (gpy_vector_t * const, int, const char *, unsigned int);
+extern void * gpy_vec_index_diag (gpy_vector_t * const, size_t, const char *, unsigned int);
 
 extern gpy_hashval_t gpy_dd_hash_string (const char *);
 extern gpy_hash_entry_t * gpy_dd_hash_lookup_table (gpy_hash_tab_t *, gpy_hashval_t);
